Move command enum to correct directory
