--- src/main/java/duke/data/task/CommandType.java
@@ -1,82 +1,0 @@
-package duke.data.task;
-
-/**
- * Valid commands.
- */
-public enum CommandType {
-    /**
-     * Bye command.
-     */
-    BYE("bye"),
-
-    /**
-     * List command.
-     */
-    LIST("list"),
-
-    /**
-     * Mark command.
-     */
-    MARK("mark"),
-
-    /**
-     * Unmark command.
-     */
-    UNMARK("unmark"),
-
-    /**
-     * Todo command.
-     */
-    TODO("todo"),
-
-    /**
-     * Deadline command.
-     */
-    DEADLINE("deadline"),
-
-    /**
-     * Event command.
-     */
-    EVENT("event"),
-
-    /**
-     * Delete command.
-     */
-    DELETE("delete");
-
-    /**
-     * Command string.
-     */
-    private final String command;
-
-    /**
-     * Constructor for the Command enum.
-     *
-     * @param command command string
-     */
-    CommandType(String command) {
-        this.command = command;
-    }
-
-    /**
-     * Retrieve type
-     */
-    public String getCommand() {
-        return this.command;
-    }
-
-    /**
-     * Checks that the given command is a valid command.
-     *
-     * @param command command to be validated
-     * @return boolean, whether the command is valid or not
-     */
-    public static boolean isValidCommand(String command) {
-        for (CommandType e : values()) {
-            if (e.command.equals(command)) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
